requirements.txt
jupyter-book==0.12.0
pandas==1.3.4
matplotlib==3.3.4
ipykernel==6.4.2
xgboost==1.5.0
scikit-learn==1.0.1
seaborn==0.11.2
ipywidgets==7.6.5
docutils==0.16
sphinxcontrib-bibtex==2.2.0
tqdm==4.31.1
fastapi==0.70.0
uvicorn==0.15.0
psycopg2-binary==2.9.2
sqlalchemy-utils==0.37.9
gunicorn==20.1.0
PyJWT==2.3.0
cryptography==36.0.1
alembic==1.7.5
ghp-import==2.0.2